rework (read): Update verilogSynopsys for compile_ultra
compile_ultra uses design ware components that required a rework to
verilogSynopsys(). FADDX1 has both 'CO'and 'S' as outputs, while MUX21X1
has 'S' as an input. To avoid future conflicts, the function was reworked
to have the ports explicately detailed for each gate type.

A FADDX1 gate will create 2 FADDX1 type nodes in the returned graph.
This is consistant with the assumption of port-agnostic for inputs.
This means converting to a graph loses information.