Add self-loops to nodes connected to PO
This is done in Lilas and circumvents the problem of a node only connected to
PIs and POs. By previous code, this node would have no edges in the npz matrix
used by the GNN and would therefore never be sampled.

Also remove use of subgraph because the nodelist parameter of
nx.convert_matrix.to_scipy_sparse_matrix will filter the output matrix to only
nodes in the nodelist.